refactor(app): remove next.config
