+++ pages/_document.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react';
-import Document, {Html, Head, Main, NextScript } from 'next/document';
+import Document, { Html, Head, Main, NextScript } from 'next/document';
 
 import { GA_TRACKING_ID } from '../core/gtag';
